+++ R/pimage.R
@@ -279,7 +279,7 @@ pimage.matrix <-
     diag = TRUE,
     row_labels = NULL,
     col_labels = NULL,
-    prop = FALSE,
+    prop = isSymmetric(x),
     flip_axes = FALSE,
     reverse_columns = FALSE,
     ...,