+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New Features
 - permute for dendrograms gained parameter dist and accepts now seriation 
   methods.
+- Added method "AOE" for correlation matrices.
 
 # seriation 1.5.1 (07/20/2023)